Seperated Button Group and Header into different components
