+++ App.js
@@ -1,89 +1,22 @@
 import React, { Component } from 'react';
-import { Platform, StyleSheet, Text, View, TouchableHighlight } from 'react-native';
+import { StyleSheet, View } from 'react-native';
+import FitHeader from './components/FitHeader';
+import FitButtons from './components/FitButtons';
 import FitMap from './components/FitMap';
 
-
-import { Header, ButtonGroup, Button, Avatar, List, ListItem, Card, Input, Icon } from 'react-native-elements';
-
-const instructions = Platform.select({
-    ios: 'Press Cmd+R to reload,\n' +
-      'Cmd+D or shake for dev menu',
-    android: 'Double tap R on your keyboard to reload,\n' +
-      'Shake or press menu button for dev menu',
-  });
-
-type Props = {};
+//type Props = {};
 export default class App extends Component<Props> {
-constructor () {
-  super()
-  this.state = {
-    selectedIndex: 2
-  }
-  this.updateIndex = this.updateIndex.bind(this)
-}
-updateIndex (selectedIndex) {
-  this.setState({selectedIndex})
-}
   render() {
-  const buttons = ['Home', 'Past Quest', 'Next Quest', 'Stats']
-  const { selectedIndex } = this.state
-  const list = [
-  {
-    title: 'Steps',
-    value: '2,867'
-  },
-  {
-    title: 'Distance',
-    value: '1 mile'
-  },
-  {
-    title: 'Elevation',
-    value: '1 mile'
-  },
-  {
-    title: 'Time',
-    value: '30 mins'
-  },
-  {
-    title: 'Avg Speed',
-    value: '10 mph'
-  }
-  
-]
     return (
       <View style={styles.container}>
-         <Card containerStyle={{flexDirection: 'row', width: "100%",
-            height: 80, backgroundColor: 'rgba(49, 111,244, 1)',
-            borderColor: 'rgba(49, 111,244, 1)', justifyContent: 'space-around',
-            alignItems: 'stretch', marginTop: 0
-            }}> 
-              <Header
-                  statusBarProps={{ barStyle: 'light-content' }}
-                  centerComponent={{ text: 'FitQuest', style: { color: '#fff',
-                  fontWeight: 'bold', fontSize: 20, marginTop: 10 } }}
-                  outerContainerStyles={{ backgroundColor: 'rgba(49, 111,244, 1)',
-                   flexDirection: 'row', borderColor: 'rgba(49, 111,244, 1)',
-                   width: '100%', justifyContent: 'space-around', alignItems: 'center' }}
-                  innerContainerStyles={{ justifyContent: 'space-between' }}/>
-              
-            </Card>
-            
-                <ButtonGroup
-                  onPress={this.updateIndex}
-                  selectedIndex={selectedIndex}
-                  buttons={buttons}
-                  buttonStyle={{justifyContent: 'center', width: '100%'}}
-                  containerStyle={{height: 70, width: '100%',
-                  backgroundColor: 'rgba(49, 111,244, 1)',
-                  justifyContent: 'flex-start', marginTop: 0}}
-                  textStyle={{color: 'white', fontWeight: 'bold'}}
-                />
-  
+        <FitHeader />
+        <FitButtons />
         <FitMap />
       </View>
     );
   }
 }
+
 const styles = StyleSheet.create({
   container: {
     flex: 1,
